+++ server/src/part_matching/requirements.txt
@@ -12,7 +12,6 @@ nltk==3.6.2
 numpy==1.19.5
 pandas==1.1.5
 Pillow==8.2.0
-pkg-resources==0.0.0
 pyparsing==2.4.7
 python-dateutil==2.8.1
 pytz==2021.1
